reduxを導入しようとしてる
indexとmainの分離
