+++ src/redux/store/index.jsx
@@ -1,0 +1,7 @@
+import { createStore } from 'redux';
+import reducer from '../reducers/';
+
+export default function configureStore() {
+	const store = createStore(reducer);
+	return store;
+}
